+++ framework/model/src/main/java/is/codion/framework/model/EntityEditModel.java
@@ -83,11 +83,6 @@ public interface EntityEditModel {
 	EntityEditor editor();
 
 	/**
-	 * @return the {@link EntityPersistence} used by this edit model
-	 */
-	EntityPersistence persistence();
-
-	/**
 	 * @return the {@link PersistTasks}
 	 */
 	PersistTasks tasks();